Migrate to go1.16
- fixes for tests
- graceful for go1.15 and below
- graceful for go1.16 and above
