+++ grace/grace_test.go
@@ -7,23 +7,21 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"go.uber.org/zap"
+	"go.uber.org/zap/zaptest"
 )
 
 func TestGrace(t *testing.T) {
-	var (
-		log = zap.L()
-		ctx = NewGracefulContext(log)
-	)
-
-	// waiting to run the goroutine and channel of signals
-	<-time.Tick(100 * time.Millisecond)
-
 	signals := []syscall.Signal{syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP}
 	for i := range signals {
 		sig := signals[i]
 		t.Run(fmt.Sprintf("should cancel context on %s signal", sig), func(t *testing.T) {
 			is := assert.New(t)
+
+			log := zaptest.NewLogger(t)
+			ctx := NewGracefulContext(log)
+
+			// waiting to run the goroutine and channel of signals
+			<-time.After(time.Millisecond)
 
 			err := syscall.Kill(syscall.Getpid(), sig)
 			is.NoError(err)
